+++ react/__mocks__/vtex.delivery-promise-components/DeliveryPromiseContext.js
@@ -1,0 +1,2 @@
+export const useDeliveryPromiseState = () => ({})
+export const useDeliveryPromiseDispatch = () => () => {}