+++ react/__mocks__/vtex.session-client.js
@@ -1,0 +1,10 @@
+export const useRenderSession = () => ({
+  session: {
+    namespaces: {
+      store: {
+        channel: { value: '1' },
+      },
+    },
+  },
+  loading: false,
+})