+++ react/constants/pickupSearch.js
@@ -1,0 +1,2 @@
+/** localStorage key for PLP pickup selection (search-result FilterNavigator). */
+export const PICKUP_IN_POINT_STORAGE_KEY = 'vtex.search.pickupInPoint'